Backend/requirements.txt
# Core API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# Gmail API
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Supabase
supabase>=2.4.0

# ML — pre-trained models
scikit-learn>=1.4.0       # SVM + Naive Bayes (.joblib)
joblib>=1.4.0             # loading .joblib files
numpy>=1.26.0
tensorflow>=2.16.0        # LSTM (.keras)

# Utilities
python-dotenv>=1.0.0